Add SmartContract controller
