+++ apps/block_scout_web/lib/block_scout_web/api_router.ex
@@ -28,7 +28,6 @@ defmodule BlockScoutWeb.ApiRouter do
   pipeline :api_v2 do
     plug(CheckApiV2)
     plug(:fetch_session)
-    plug(:protect_from_forgery)
   end
 
   alias BlockScoutWeb.Account.Api.V1.{TagsController, UserController}
@@ -130,12 +129,13 @@ defmodule BlockScoutWeb.ApiRouter do
     end
 
     scope "/smart-contracts" do
-      get("/:address_hash", V2.AddressController, :smart_contract)
-      get("/:address_hash/methods-read", V2.AddressController, :methods_read)
-      get("/:address_hash/methods-write", V2.AddressController, :methods_write)
-      get("/:address_hash/methods-read-proxy", V2.AddressController, :methods_read_proxy)
-      get("/:address_hash/methods-write-proxy", V2.AddressController, :methods_write_proxy)
-      get("/:address_hash/query-read-method", V2.AddressController, :query_read_method)
+      get("/:address_hash", V2.SmartContractController, :smart_contract)
+      get("/:address_hash/methods-read", V2.SmartContractController, :methods_read)
+      get("/:address_hash/methods-write", V2.SmartContractController, :methods_write)
+      get("/:address_hash/methods-read-proxy", V2.SmartContractController, :methods_read_proxy)
+      get("/:address_hash/methods-write-proxy", V2.SmartContractController, :methods_write_proxy)
+      get("/:address_hash/query-read-method", V2.SmartContractController, :query_read_method)
+      post("/:address_hash/query-read-method", V2.SmartContractController, :query_read_method)
     end
 
     scope "/tokens" do
